JiKong-Modbus-BMS-PB2A16S30P-addon/requirements.txt
paho-mqtt==2.1.0
pymodbus==3.5.0
PyYAML==6.0.1
schedule
astral
pytz
jq
